\name{osmose2R-package}
\alias{osmose2R-package}
\alias{osmose2R}
\docType{package}
\title{Analize OSMOSE model outputs within R}
\description{
Methods to analize and plot OSMOSE model outputs.}
\details{
\tabular{ll}{
Package: \tab osmose2R\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-03-07\cr
License: \tab TBD\cr
}

The function \emph{osmose2R} get OSMOSE outputs into R.
}
\author{
Ricardo Oliveros-Ramos

Maintainer: Ricardo Oliveros-Ramos <ricardo.oliveros@gmail.com>
}
\references{
Not yet published.
}
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
\dontrun{
osmose2R()
}
}
